+++ src/services/HttpRequest.ts
@@ -1,0 +1,8 @@
+import { Context } from 'effect';
+
+export interface RequestContext {
+	req: Request;
+	resHeaders: Headers;
+}
+
+export class HttpRequest extends Context.Tag('HttpRequest')<HttpRequest, RequestContext>() {}